Fix offsets for mult
Previously, we assumed that the a_stripe and b_stripe would need to have
an offset relative to location in the integrated array. Since we are
only passing stripes, we have a fixed location relative to index 0 for
all stripes.

For the c_stripe, we simply shift left and right along the columns to
fill out all columns of the strip. Rows are constant throughout and are
0 indexed like a_stripe is.

I could be very wrong here still...just documenting along the way.